Implement tile type logic for bullets
