+++ packages/common/src/collisions/types.ts
@@ -1,0 +1,2 @@
+export const COLLISION_TYPES = ['full', 'half'];
+export type CollisionType = 'full' | 'half';
